Update hello-world.mdx
I added additional text with links that might help devs that go right to the examples instead of reading the complete documentation immediately.  It would point them in the right direction for additional documentation should they be confused about what's going on.  Just suggestions on a more robust explanation for those of us just starting with Soroban,

These are incomplete and need review by someone more in-tune with the current environment than I am.  A few of the suggestions have xxx in them!!