don't call it a schema
No other documentation is calling it a "schema", and it doesn't seem
that this term makes sense to other people.

After discussing other alternatives, it seems like people liked the idea
of "API", but this term feels too overloaded to me. An API feels like
a service to which I can make HTTP calls in order to get JSON or some
other machine-readable format. But the thing we're describing here is
_how_ to call such a service. That is, if you think of each contract as
such an API service, then the thing we want to describe is the
machine-readable description of how to call it.

So here I'm trying out calling the page "Fully-Typed Contracts", and
then using the maybe-too-verbose "interface types" throughout the
document, rather than either "schema" or "API".