fix: move State Expiration to Concepts
This is not part of the core onboarding tutorial. It probably belongs in
Fundamentals and Concepts instead.

For background, these docs are attempting to use [Divio's documentation
system](https://documentation.divio.com/), with the "Getting Started"
section as _The_ Tutorial. The "Fundamentals and Concepts" section is
what Divio describes as "Explanation" docs.