feat: move Freighter and Running RPC to Reference
We want to keep the stuff in Getting Started hyper-focused on a quick
end-to-end tour of Soroban. Now that there's a public RPC endpoint, it
is much more straightforward to have people deploy contract right to it,
rather than running their own nodes. Running your own node is a more
advanced concept that dabblers need not muck with.

So this moves both the "Deploy to a Local Network" and "Deploy to
Futurenet" pages to the RPC page, in the Reference section. This
required some reorganization & rework of the concepts to make it fit the
new context. I also updated some instructions, such as teaching
people to use `soroban config network` and `soroban config identity`,
which were not included in the old "Deploy to…" pages.

This commit also moves the "Connect Freighter Wallet" to Reference,
since it's a bit in-the-weeds, which is a good feature for a Reference
document but not for an end-to-end Tutorial.

Coming Soon™: a new "Deploy Contracts" step in the Getting Started
section, which uses the public RPC Futurenet endpoint. This will be
followed by a "Create a Dapp" step, which will use the new JS Lib
generation built into the CLI, and go over relevant Freighter details.