feat: Display example RPC requests as a useable CURL command
Also displaying generic code snippets for popular languages, as
well as in raw JSON format.

A new `ExampleRequest` component has been made to facilitate this,
as well. This component takes a `method` prop for the name of the
RPC method being called, as well as an optional `params` prop which
can pass the object that should be used to populate the RPC request
object.

Refs: #565